requirements.txt
beautifulsoup4==4.3.2
colorama==0.3.3
dj-database-url==0.3.0
dj-static==0.0.6
Django==1.8.5
django-braces==1.8.1
django-cors-headers==1.1.0
django-countries==3.3
django-filter==0.10.0
django-mssql==1.6.2
django-oauth-plus==2.2.6
django-oauth-toolkit==0.8.2
django-oauth2-provider==0.2.6.1
django-postgrespool==0.3.0
django-rest-framework-social-oauth2==0.0.8
django-rest-swagger==0.3.2
django-sqlserver==1.7
django-toolbelt==0.0.1
django-wpadmin==1.7.4
djangorestframework==3.1.3
djangorestframework-oauth==1.0.1
djoser==0.2.1
gitdb==0.6.0
GitPython==0.3.2.1
gunicorn==19.3.0
httpie==0.9.2
httplib2==0.9.1
ipython==3.2.0
Markdown==2.6.2
oauth2==1.5.211
oauthlib==1.0.0
Pillow==2.9.0
psycopg2==2.6.1
Pygments==2.0.2
PyJWT==1.4.0
pyserial==2.7
PySide==1.2.2
python-openid==2.2.5
python-social-auth==0.2.12
pyusb==1.0.0b1
PyYAML==3.11
requests==2.7.0
requests-oauthlib==0.5.0
shortuuid==0.4.2
six==1.9.0
smmap==0.8.3
South==1.0.2
sql-server.pyodbc==1.0
SQLAlchemy==1.0.9
static3==0.6.1
virtualenv==13.0.1
wheel==0.24.0
whitenoise==1.0.6
